refactor(gulmc): resolve coverage dependency per item via source_item_id
Follows the review of #2063. The correlations file now carries source_item_id in
place of source_coverage_id (same u4, record unchanged at 24 bytes): each
dependent item names the source item at its own location, building and peril,
and gulmc derives the coverage-level dependency forest from those item links.

Why: the engine previously inferred a dependent item's counterpart from its
position after items are sorted by (areaperil_id, vulnerability_id). Within one
areaperil that tiebreak is the vulnerability id, and a source's hazard-indexed
ids and a dependent's conditional ids come from different id spaces, so their
orders need not agree. With two perils geocoded to the same cell the perils were
silently crossed — a dependent driven by the wrong peril's source damage. Naming
the source item removes the inference entirely, and retires the areaperil-multiset
guard that existed only to make positional pairing safe.

Resolution is now per item rather than per coverage, which makes mixed
vulnerability kinds on one coverage type work: an item that finds no source item
(no source coverage at the location, or the configured pair returned at different
areaperils) is computed independently, and gulmc — the only stage that knows
which vulnerability ids are conditional — refuses it only if its vulnerability is
a conditional one. So a model may supply a conditional vulnerability where the
cells align and a hazard-indexed one where they do not.

Zero-tiv source retention is tested on (loc_id, peril_id, areaperil_id), the same
triple a dependent pairs on, so nothing is retained that cannot drive a dependent.

conditional_vulnerability is registered with the standard converters as the
conditionalvulnerability file type, with its own dtype whose field names match the
file (source_damage_bin, damage_bin) and a binary layout interchangeable with a
flat vulnerability file. Conversion validation is shared with vulnerability and
gains a duplicate check: damage bins must strictly increase within a group, since
both loaders scatter by assignment and would otherwise silently keep only the last
row of a repeated triple, leaving the column short of 1. That column would then
sample past the top of its last defined damage bin — extrapolating, not clamping,
so a loss can exceed the coverage's tiv. The loader no longer re-checks column
sums; that lives with the equivalent vulnerability checks in the converter.

Also fixes a pre-existing logging leak: logging_reset_handlers restored the
handlers of every 'oasislmf.*' logger but left their level pinned at the run's log
level, and restored propagate only when the logger happened to have a handler to
iterate over. INFO and DEBUG from every oasislmf submodule was therefore
suppressed for the rest of any process that ran gulmc or gulpy — including
oasislmf model run, which continues with output steps after the engine.

tests/assets/test_model_8 now mixes damage group ids: coverages 1 and 2 share one
(as the default damage_group_id_cols produce, since they do not include coverage
type, which rank-couples a source and its dependent) while coverage 3 has its own.
Both the coupled and uncoupled sampling cases are therefore covered by the golden
files; 108 of the 144 changed, the S=0 cells having no samples.

2212 tests pass across tests/pytools, tests/preparation and tests/computation.

+++ tests/pytools/gulmc/test_rp_protection.py
@@ -518,7 +518,8 @@ def test_rp_protection_only_affects_protected_items():
     vuln_array = np.zeros((1, Ndamage_bins, Nintensity_bins), dtype=oasis_float)
     vuln_array[0, 1, 0] = 1.0
     conditional_vuln_array = np.zeros((0, Ndamage_bins, Ndamage_bins), dtype=oasis_float)
-    vuln_idx_to_cond_idx = np.full(vuln_array.shape[0], -1, dtype=np.int32)
+    # int64 as build_structures produces it, so this exercises the specialisation the engine compiles
+    vuln_idx_to_cond_idx = np.full(vuln_array.shape[0], -1, dtype=np.int64)
 
     damage_bins = np.zeros(Ndamage_bins, dtype=damagebin_dtype)
     damage_bins[0] = (0, 0.0, 0.5, 0.25, 0)
